fix bundled imports: fix incorrect hashed turbo import
